Module 2

+++ src/main/java/com/example/ecommerce/Service/UserService.java
@@ -1,14 +1,19 @@
 package com.example.ecommerce.Service;
 
+import com.example.ecommerce.Advices.ExceptionHandling.CustomExceptions.ResourceNotFoundException;
 import com.example.ecommerce.DTO.UserDto;
 import com.example.ecommerce.Entity.User;
 import com.example.ecommerce.Repository.UserRepository;
 import lombok.RequiredArgsConstructor;
 import org.modelmapper.ModelMapper;
+import org.springframework.http.ResponseEntity;
 import org.springframework.stereotype.Service;
+import org.springframework.util.ReflectionUtils;
 
+import java.lang.reflect.Field;
 import java.util.List;
-import java.util.stream.Collectors;
+import java.util.Map;
+
 
 @Service
 @RequiredArgsConstructor
@@ -18,27 +23,62 @@ public class UserService {
     private final ModelMapper modelMapper;
 
 
-    public Boolean deleteById(Long id) {
-         userRepository.deleteById(id);
-         return true;
+    public ResponseEntity<Boolean> deleteById(Long id) {
+        boolean isExist = userRepository.existsById(id);
+        if(!isExist)
+                throw new ResourceNotFoundException(" resource with id " + id + " does not exist");
+        userRepository.deleteById(id);
+         return ResponseEntity.ok(Boolean.TRUE);
     }
 
-    public List<UserDto> findAllUsers() {
-        return userRepository.findAll()
+    public ResponseEntity<List<UserDto>> findAllUsers() {
+         List<UserDto> userDtoList =   userRepository.findAll()
                 .stream()
                 .map(user -> modelMapper.map(user, UserDto.class))
-                .collect(Collectors.toList());
+                .toList();
+         return ResponseEntity.ok(userDtoList);
     }
 
-    public UserDto findById(Long id) {
-        return userRepository.findById(id).
-                map(user -> modelMapper.map(user , UserDto.class))
-                .orElseThrow();
+    public ResponseEntity<UserDto> findById(Long id) {
+        User user = userRepository.findById(id).orElse(null);
+        if(user == null)
+            throw new ResourceNotFoundException(" resource with id " + id + " does not exist");
+        return ResponseEntity.ok(modelMapper.map(user, UserDto.class));
     }
 
-    public UserDto createUser(UserDto dto) {
+    public ResponseEntity<UserDto> createUser(UserDto dto) {
         User user = modelMapper.map(dto , User.class);
         User saved = userRepository.save(user);
-        return modelMapper.map(saved , UserDto.class);
+        return ResponseEntity.ok(modelMapper.map(saved, UserDto.class));
+    }
+
+    public ResponseEntity<UserDto> updateUserById(Long id, UserDto userDto) throws Exception {
+        boolean isExist = userRepository.existsById(id);
+        if(!isExist)
+            throw new ResourceNotFoundException(" resource with id " + id + " does not exist");
+
+
+        User user = modelMapper.map(userDto , User.class);
+        UserDto dto = modelMapper.map( userRepository.save(user) , UserDto.class);
+        return ResponseEntity.ok(dto);
+    }
+
+
+    public ResponseEntity<UserDto> partiallyUpdateUserById(Long id, Map<String, Object> map) throws Exception {
+
+        boolean isExist = userRepository.existsById(id);
+        if(!isExist)
+            throw new ResourceNotFoundException(" resource with id " + id + " does not exist");
+
+        User user = userRepository.findById(id).orElseThrow();
+
+        map.forEach((key, value) -> {
+            Field field = ReflectionUtils.findField(User.class , key);
+            field.setAccessible(true);
+            ReflectionUtils.setField(field,user,value);
+        });
+        User savedUser = userRepository.save(user);
+        return ResponseEntity.ok(modelMapper.map(savedUser,UserDto.class));
+
     }
 }
